functions/requirements.txt
azure-functions
httpx>=0.27.0
python-dotenv>=1.0.0
azure-storage-blob>=12.19.0
